requirements.txt
python-digitalocean==1.8.1
fabric==1.11.1
